% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_with_ci.R
\name{barplot_with_interval}
\alias{barplot_with_interval}
\title{Barplot with Intervals}
\usage{
barplot_with_interval(x, method = "minmax", ylim = NULL, ...)
}
\arguments{
\item{x}{matrix}

\item{method}{method for intervals, so far only "minmax"}

\item{ylim}{limits for y axis. Default:NULL is ylim=c(0,max(interval))}

\item{...}{additional parameters forwarded to barplot}
}
\value{
plot
}
\description{
Barplot with Intervals
}

