% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_with_ci.R
\name{barplot_with_interval_23}
\alias{barplot_with_interval_23}
\title{Barplot with Intervals for two or three bars beside}
\usage{
barplot_with_interval_23(x, l, method = "minmax", ylim = NULL, ...)
}
\arguments{
\item{x}{array}

\item{l}{number of bars beside (second dimension of x)}

\item{method}{method for intervals, so far only "minmax"}

\item{ylim}{limits for y axis. Default:NULL is ylim=c(0,max(interval))}

\item{...}{additional parameters forwarded to barplot}
}
\value{
plot
}
\description{
Barplot with Intervals for two or three bars beside
}

