% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_neighbours.R
\name{class.neighbours.folder}
\alias{class.neighbours.folder}
\title{class.neighbours.folder}
\usage{
class.neighbours.folder(inputfolder, outputfolder, N = 7)
}
\arguments{
\item{inputfolder}{Input folder}

\item{outputfolder}{Output folder}

\item{N}{Max class
#'}
}
\value{
plots
}
\description{
class.neighbours.folder
}
