% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{classify.folder}
\alias{classify.folder}
\title{Classify DAPI}
\usage{
classify.folder(f, N, beta = 0.1, output = paste0("class", N), cores = 1)
}
\arguments{
\item{f}{folder}

\item{N}{number of classes}

\item{beta}{beta parameter used in bioimagetools::segment()}

\item{output}{output folder}

\item{cores}{number of cores used in parallel (needs parallel package)}
}
\value{
results in "output" and "output"-n
}
\description{
Classify DAPI
}

