% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R
\name{classify}
\alias{classify}
\title{Classify DAPI}
\usage{
classify(f, N, cores = 1)
}
\arguments{
\item{f}{folder}

\item{N}{number of classes}

\item{cores}{number of cores used in parallel (needs parallel package)}
}
\value{
results in classN/ and classN-n
}
\description{
Classify DAPI
}

