% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{classify.single}
\alias{classify.single}
\title{Classify DAPI}
\usage{
classify.single(...)
}
\arguments{
\item{...}{parameters for classify}
}
\value{
image with classes
}
\description{
These functions are provided for compatibility with older version of
the nucim package.  They may eventually be completely
removed.
}

