% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dapimask_file.R
\name{dapimask.file}
\alias{dapimask.file}
\title{Automatic DAPI mask segmentation for files}
\usage{
dapimask.file(file, folder = "blue", pixelsize = NULL, size = NULL,
  cores = 1)
}
\arguments{
\item{file}{file to read}

\item{folder}{with}

\item{pixelsize}{real pixelsize of image (in nanometers), if NULL (default), look in folder XYZmic}

\item{size}{real size of image (in microns), if NULL (dafault), look in folder XYZmic}

\item{cores}{Number of cores available for parallel computing}
}
\value{
nothing, DAPI mask image will be saved to dapimask/
}
\description{
Automatic DAPI mask segmentation for files
}

