% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract_spots_DHS.R
\name{extract.spots.DHS}
\alias{extract.spots.DHS}
\title{Title Extract spots for DHS FISH}
\usage{
extract.spots.DHS(file, folder = "./", thresh.offset = 0.1,
  min.sum.intensity = 2, max.distance = 0.5, use.brightest = FALSE,
  full.voxels = FALSE, output = "markers")
}
\arguments{
\item{file}{File name}

\item{folder}{Folder}

\item{thresh.offset}{Thresh offest used in EBImage::thresh()}

\item{min.sum.intensity}{spots smaller than min.sum.intensity are ignored}

\item{max.distance}{use only spots with distance to other color spot smaller than max.distance}

\item{use.brightest}{Logical; use only brightest in max.distance?}

\item{full.voxels}{Logical; output contains full voxels instead of rgb intensities}

\item{output}{output folder}
}
\value{
RGB image with spots will be written to output folder
}
\description{
Title Extract spots for DHS FISH
}

