% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_spots.R
\name{extract.spots}
\alias{extract.spots}
\title{Extract spots using adaptive thresholding and minimum spot intensitiy}
\usage{
extract.spots(file, folder = "./", thresh.offset = 0.1,
  min.sum.intensity = 2, full.voxels = FALSE, output = "markers")
}
\arguments{
\item{file}{name of file, this file needs to be in red/, green/, class7/ and dapimask/}

\item{folder}{name of root folder}

\item{thresh.offset}{thresholf value for adaptive thresholding}

\item{min.sum.intensity}{threshold for spot sum intensity}

\item{full.voxels}{if TRUE returen 0/1, if FALSE returns spots with intensitites}

\item{output}{output folder name, output will be an rgb stack with DAPI class and red/green spots}
}
\description{
Extract spots using adaptive thresholding and minimum spot intensitiy
}

