% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_spots.R
\name{find.spots.file}
\alias{find.spots.file}
\title{Detects spotsfor one file}
\usage{
find.spots.file(file, dir, color, thresh = NULL, thresh.auto = FALSE,
  thresh.quantile = 0.9, filter = NULL, cores = 1)
}
\arguments{
\item{file}{file}

\item{dir}{directory for results}

\item{color}{which color, images have to be in folder with color name}

\item{thresh}{threshold}

\item{thresh.auto}{Logical. Find threshold automatically?}

\item{thresh.quantile}{numeric. use simple}

\item{filter}{2d-filter to use before spot detection}

\item{cores}{number of cores to use in parallel (with parallel package only)}
}
\value{
spot images in spot-color/, number of spots as txt files in spot-color/
}
\description{
Detects spotsfor one file
}

