% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_spots.R
\name{find.spots.folder}
\alias{find.spots.folder}
\title{Detects spots}
\usage{
find.spots.folder(
  f,
  color,
  thresh = 1,
  thresh.auto = TRUE,
  filter = NULL,
  cores = 1
)
}
\arguments{
\item{f}{path to folder}

\item{color}{which color, images have to be in folder with color name}

\item{thresh}{threshold}

\item{thresh.auto}{Logical. Find threshold automatically?}

\item{filter}{2d-filter to use before spot detection}

\item{cores}{number of cores to use in parallel (with parallel package only)}
}
\value{
spot images in spot-color/, number of spots as txt files in spot-color/
}
\description{
Detects spots
}
