% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap7.R
\name{heatmap.color}
\alias{heatmap.color}
\title{Heatmap colors for n classes}
\usage{
heatmap.color(n)
}
\arguments{
\item{n}{number of colors.}
}
\description{
Heatmap colors for n classes
}
\examples{
barplot(8:1,col=heatmap.color(8))
}
