% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap7.R
\name{heatmap7}
\alias{heatmap7}
\title{Heatmap colors for 7 classes}
\usage{
heatmap7(...)
}
\arguments{
\item{...}{parameters are ignored.}
}
\description{
Heatmap colors for 7 classes
}
\examples{
barplot(7:1,col=heatmap7())
}

