% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dapimask_file.R
\name{mask.dapi.file}
\alias{mask.dapi.file}
\title{Automatic DAPI mask segmentation for files}
\usage{
mask.dapi.file(file, folder = "blue")
}
\arguments{
\item{file}{file to read}

\item{folder}{with}
}
\value{
nothing, DAPI mask image will be saved to dapimask/
}
\description{
Automatic DAPI mask segmentation for files
}

