% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dapimask_folder.R
\name{mask.dapi.folder}
\alias{mask.dapi.folder}
\title{automatic DAPI mask segmentation for folder}
\usage{
mask.dapi.folder(path, folder = "blue", cores = 1)
}
\arguments{
\item{path}{path to folder with DAPI}

\item{folder}{folder with DAPI images}

\item{cores}{number of cores to use in parallel (need paralle package)}
}
\value{
nothing, results are in folder dapimask
}
\description{
automatic DAPI mask segmentation for folder
}

