% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mask_dapi.R
\name{mask.dapi}
\alias{mask.dapi}
\title{Mask DAPI in kernel}
\usage{
mask.dapi(img, mic, thresh = "auto")
}
\arguments{
\item{img}{DAPI channel image (3d)}

\item{mic}{vector of dimensions of img in microns}

\item{thresh}{threshold for intensity. Can be "auto": function will try to find automatic threshold}
}
\value{
mask image, same dimension as img.
}
\description{
Mask DAPI in kernel
}

