% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_class_folder.R
\name{nearestClassDistances.folder}
\alias{nearestClassDistances.folder}
\title{Find all distances to next neighbour of all classes for folders}
\usage{
nearestClassDistances.folder(path, N = 7, voxelsize = NULL, cores = 1)
}
\arguments{
\item{path}{path to folder}

\item{N}{number of classes, default: 7}

\item{voxelsize}{real voxesize of image (in nanometers), if NULL (default), look in folder XYZmic}

\item{cores}{number of cores to use in parallel (needs parallel package)}
}
\value{
nothing, results are in folder distances in RData format
}
\description{
Find all distances to next neighbour of all classes for folders
}

