% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/splitchannels.R
\name{classify.single}
\alias{classify.single}
\alias{splitchannels}
\title{Classify DAPI}
\usage{
classify.single(...)

splitchannels(...)
}
\arguments{
\item{...}{parameters for classify}

\item{...}{parameters for splitchannels.folder}
}
\value{
image with classes

Nothing, folders red, green, blue and XYZmic include seperate channels and pixel size information
}
\description{
These functions are provided for compatibility with older version of
the yourPackageName package.  They may eventually be completely
removed.

These functions are provided for compatibility with older version of
the yourPackageName package.  They may eventually be completely
removed.
}

