% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_classify_folder.R
\name{plot_classify.folder}
\alias{plot_classify.folder}
\title{Plot barplot for classified images in a folder}
\usage{
plot_classify.folder(path, N = 7, cores = 1,
  col = grDevices::grey(0.7), method = "sd")
}
\arguments{
\item{path}{path to folder}

\item{N}{number of classes, default: 7}

\item{cores}{number of cores to use in parallel (needs parallel package if cores>1)}

\item{col}{color of bars, either one or a vector of hex RGB characters}

\item{method}{method for error bars ("sd", "minmax", "quartile")}
}
\value{
plots
}
\description{
Plot barplot for classified images in a folder
}
