% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_colors_in_classes.R
\name{plot_colors.in.classes.folder}
\alias{plot_colors.in.classes.folder}
\title{Plot for colors in classes distribution for folders}
\usage{
plot_colors.in.classes.folder(path, col1 = "green", col2 = "red")
}
\arguments{
\item{path}{path to folder}

\item{col1}{color of channel 1}

\item{col2}{color of channel 2}
}
\value{
plot
}
\description{
Plot for colors in classes distribution for folders
}
