% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_classify_folder.R
\name{plotclassify.folder}
\alias{plotclassify.folder}
\title{Plot barplot for classified images in a folder}
\usage{
plotclassify.folder(path, N = 7, cores = 1)
}
\arguments{
\item{path}{path to folder}

\item{N}{number of classes, default: 7}

\item{cores}{number of cores to use in parallel (needs parallel package if cores>1)}
}
\value{
plots
}
\description{
Plot barplot for classified images in a folder
}

