% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rgb-split.R
\name{rgbsplit}
\alias{rgbsplit}
\title{Split RGB images into channels and pixel size information}
\usage{
rgbsplit(path, channels = c("red", "green", "blue"), rgb.folder = "rgb",
  cores = 1)
}
\arguments{
\item{path}{Path to root folder}

\item{channels}{Vector of channels in images}

\item{rgb.folder}{Folder with RGB images}

\item{cores}{Number of cores used in parallel, cores=1 implies no parallelization}
}
\value{
Nothing, folders red, green, blue and XYZmic include seperate channels and pixel size information
}
\description{
Split RGB images into channels and pixel size information
}
\examples{
rgb.split("./")
}

