% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/split_channel.R
\name{split.channel}
\alias{split.channel}
\title{Split RGB channels}
\usage{
\method{split}{channel}(img, preprocess = TRUE)
}
\arguments{
\item{img}{}
}
\value{
list with red, green, blue channels and size in microns.
}
\description{
Split RGB channels
}

