% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rgb-split.R
\name{split.channels.file}
\alias{split.channels.file}
\title{Split channels into files and extracts size in microns}
\usage{
\method{split}{channels.file}(file, channels, rgb.folder)
}
\arguments{
\item{file}{file name}

\item{channels}{e.g. c("red","green","blue")}

\item{rgb.folder}{folder with file}
}
\value{
files in "./red/", "./green", "./blue" and "./XYZmic"
}
\description{
Split channels into files and extracts size in microns
}

