% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitchannel.R
\name{splitchannel}
\alias{splitchannel}
\title{Split RGB channels}
\usage{
splitchannel(img, preprocess = TRUE)
}
\arguments{
\item{img}{rgb image}

\item{preprocess}{logical. Should preprocessing be applied?}
}
\value{
list with red, green, blue channels and size in microns.
}
\description{
Split RGB channels
}
