% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitchannels.R
\name{splitchannels.file}
\alias{splitchannels.file}
\title{Split channels into files and extracts size in microns}
\usage{
splitchannels.file(file, channels, rgb.folder, normalize = FALSE)
}
\arguments{
\item{file}{file name}

\item{channels}{e.g. c("red","green","blue")}

\item{rgb.folder}{folder with file}

\item{normalize}{boolean. Should we try to do normalization?}
}
\value{
files in "./red/", "./green", "./blue" and "./XYZmic"
}
\description{
Split channels into files and extracts size in microns
}
