% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitchannels.R
\name{splitchannels}
\alias{splitchannels}
\title{Split RGB images into channels and pixel size information}
\usage{
splitchannels(...)
}
\arguments{
\item{...}{parameters for splitchannels.folder}
}
\value{
Nothing, folders red, green, blue and XYZmic include seperate channels and pixel size information
}
\description{
These functions are provided for compatibility with older version of
the nucim package.  They may eventually be completely
removed.
}
