% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_colors_in_classes_folder.R
\name{t_colors.in.classes.folder}
\alias{t_colors.in.classes.folder}
\title{Test for colors in classes distribution for folders}
\usage{
t_colors.in.classes.folder(path, test = "Wilcoxon")
}
\arguments{
\item{path}{path to folder}

\item{test}{"Wilcoxon", "wilcox" or "U" for Wilcoxon rank-sum (Mann-Whitney U), chisq for Chi-squared test}
}
\value{
test results
}
\description{
Test for colors in classes distribution for folders
}

