% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_colors_in_classes_folder.R
\name{test_colors.in.classes.folder}
\alias{test_colors.in.classes.folder}
\title{Test for colors in classes distribution for folders}
\usage{
test_colors.in.classes.folder(path, test = "chisq")
}
\arguments{
\item{path}{}

\item{col1}{}

\item{col2}{}
}
\value{
plot
}
\description{
Test for colors in classes distribution for folders
}

