% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_with_ci.R
\name{barplot_with_interval}
\alias{barplot_with_interval}
\title{Barplot with Intervals}
\usage{
barplot_with_interval(x, method = "minmax", qu = c(0, 1), ylim = NULL,
  horiz = FALSE, border = NA, ...)
}
\arguments{
\item{x}{matrix}

\item{method}{method for intervals: "minmax" (default), "quantile" or "sd"}

\item{qu}{vector of two quantiles for method="quantile}

\item{ylim}{limits for y axis. Default:NULL is ylim=c(0,max(interval))}

\item{horiz}{boolean: horizontal bars?}

\item{border}{border parameter forwarded to barplot, default: NA is nor border}

\item{...}{additional parameters forwarded to barplot}
}
\value{
plot
}
\description{
Barplot with Intervals
}

