% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_neighbours.R
\name{class.neighbours}
\alias{class.neighbours}
\title{Class neighbourhood distribution}
\usage{
class.neighbours(img, N, N.max = 7, cores = 1)
}
\arguments{
\item{img}{Class image}

\item{N}{which class}

\item{N.max}{maximum class (default: 7)}

\item{cores}{number of cores used in parallel (needs parallel package)}
}
\value{
vector of length N.max
}
\description{
Class neighbourhood distribution
}
