% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{classify}
\alias{classify}
\title{Classify DAPI}
\usage{
classify(blue, mask, N, beta, z = 1/3)
}
\arguments{
\item{blue}{DAPI channel (image)}

\item{mask}{mask (image)}

\item{N}{number of classes}

\item{beta}{smoothing parameter used in potts model}

\item{z}{number of cores used in parallel (needs parallel package)}
}
\value{
image with classes
}
\description{
Classify DAPI
}

