% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{classify.table}
\alias{classify.table}
\title{Classify DAPI from class image}
\usage{
classify.table(class, N)
}
\arguments{
\item{class}{classes image}

\item{N}{number of classes}
}
\value{
table with number of voxels per class
}
\description{
Classify DAPI from class image
}

