% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_distance2border.R
\name{compute.distance2border}
\alias{compute.distance2border}
\title{Compute distance to border of classes}
\usage{
compute.distance2border(f, color, N, from.spots = FALSE,
  output = "dist2border", cores = 1)
}
\arguments{
\item{f}{folder of classes images}

\item{color}{folder of color images ("spots-"color for spots images)}

\item{N}{which class}

\item{from.spots}{Logical.}

\item{output}{output folder}

\item{cores}{number of parallel cores which can be used}
}
\value{
images in output"-"color"-"N
}
\description{
Compute distance to border of classes
}
