% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dapimask.R
\name{dapimask}
\alias{dapimask}
\title{Automatic DAPI mask}
\usage{
dapimask(file, folder = "blue")
}
\arguments{
\item{file}{file to read}

\item{folder}{with}
}
\value{
nothing, DAPI mask image will be saved to dapimask/
}
\description{
Automatic DAPI mask
}

