% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dapimask.R
\name{dapimask}
\alias{dapimask}
\title{Mask DAPI in kernel}
\usage{
dapimask(img, mic, thresh = "auto", cores = 1)
}
\arguments{
\item{img}{DAPI channel image (3d)}

\item{mic}{vector of dimensions of img in microns}

\item{thresh}{threshold for intensity. Can be "auto": function will try to find automatic threshold}

\item{cores}{number of cores available for parallel computing}
}
\value{
mask image, same dimension as img.
}
\description{
Mask DAPI in kernel
}

