% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dapimask.R
\name{dapimask}
\alias{dapimask}
\title{Automatic DAPI mask segmentation}
\usage{
dapimask(blau, xymic)
}
\arguments{
\item{blau}{3d image with DAPI intensities}

\item{xymic}{voxel size in x/y-direction in microns}
}
\value{
mask
}
\description{
Automatic DAPI mask segmentation
}

