% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dapimask.R
\name{dapimask}
\alias{dapimask}
\title{Mask DAPI in kernel}
\usage{
dapimask(img, size = NULL, voxelsize = NULL, thresh = "auto",
  silent = TRUE, cores = 1)
}
\arguments{
\item{img}{DAPI channel image (3d)}

\item{size}{size of img in microns}

\item{voxelsize}{size of voxel in microns}

\item{thresh}{threshold for intensity. Can be "auto": function will try to find automatic threshold}

\item{silent}{Keep silent?}

\item{cores}{number of cores available for parallel computing}
}
\value{
mask image, array with same dimension as img.
}
\description{
Mask DAPI in kernel
}
