% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spots2.R
\name{find.spots}
\alias{find.spots}
\title{Detects spots}
\usage{
find.spots(f, color, thresh = 1, thresh.auto = TRUE, filter = NULL,
  cores = 1)
}
\arguments{
\item{f}{path to folder}

\item{color}{which color, images have to be in folder with color name}

\item{cores}{number of cores to use in parallel (with parallel package only)}
}
\value{
spot images in spot-color/, number of spots as txt files in spot-color/
}
\description{
Detects spots
}

