% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_class_folder.R
\name{nearestClassDistances.folder}
\alias{nearestClassDistances.folder}
\title{Find all distances to next neighbour of all classes for folders}
\usage{
nearestClassDistances.folder(path, N = 7, voxelsize = NULL, add = FALSE,
  cores = 1)
}
\arguments{
\item{path}{path to folder}

\item{N}{number of classes, default: 7}

\item{voxelsize}{real size of voxels (in microns), if NULL (default), look in folder XYZmic}

\item{add}{if TRUE, only process images which have not been processed before (i.e. have been added to classN)}

\item{cores}{number of cores to use in parallel (needs parallel package if cores>1)}
}
\value{
nothing, results are in folder distances in RData format
}
\description{
Find all distances to next neighbour of all classes for folders
}

