% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitchannels.R
\docType{package}
\name{nucim-deprecated}
\alias{nucim-deprecated}
\alias{nucim-deprecated-package}
\alias{splitchannels}
\title{Split RGB images into channels and pixel size information}
\usage{
splitchannels(...)
}
\arguments{
\item{path}{Path to root folder}

\item{channels}{Vector of channels in images}

\item{rgb.folder}{Folder with RGB images}

\item{cores}{Number of cores used in parallel, cores=1 implies no parallelization}
}
\value{
Nothing, folders red, green, blue and XYZmic include seperate channels and pixel size information
}
\description{
These functions are provided for compatibility with older version of
the yourPackageName package.  They may eventually be completely
removed.
}
\examples{
splitchannels.folder("./")

}

